+++ internal/payments/container.go
@@ -1,0 +1,16 @@
+package payments
+
+import (
+	"github.com/google/wire"
+	"github.com/kalilventura/vehicle-management-payment/internal/payments/domain/commands"
+	"github.com/kalilventura/vehicle-management-payment/internal/payments/infrastructure/controllers"
+	"github.com/kalilventura/vehicle-management-payment/internal/payments/infrastructure/repositories"
+)
+
+//nolint:gochecknoglobals // requirement for container
+var Container = wire.NewSet(
+	commands.Container,
+	controllers.Container,
+	repositories.Container,
+	NewModule,
+)