+++ internal/payments/domain/repositories/payments_repository.go
@@ -1,0 +1,7 @@
+package repositories
+
+import "github.com/kalilventura/vehicle-management-payment/internal/payments/domain/entities"
+
+type PaymentsRepository interface {
+	Save(payment *entities.Payment) error
+}