+++ CHANGELOG.md
@@ -17,4 +17,5 @@ Exceptions are acceptable depending on the circumstances (critical bug fixes tha
 
 ### Added
 
+- added a new resource called `POST /v1/payments` to process the vehicle payment
 - added project structure